test: use stm's retry to wait for finishing flags.
We used 'threadDelay' to wait for some time and checked the
flags ('terminated' and 'released'). However, this heavily relies on the
process/thread scheduling policy. Travis tests failed sometimes.

So now we use 'retry' to wait for the flags to satisfy the desired
condition. That way, it's more robust to scheduling uncertainty. For
safe-guard, we set 'timeout' to 20 seconds, though.